Caching the Inverse of a Matrix
This R file includes two functions, which are used to create a special object that stores a matrix and caches its inverse. The first function, makeCacheMatrix creates a special “matrix”, which is really a list containing a function to set and get the matrix and the inverse.  The second function computes the inverse of the special “matrix” returned by the first one. If the inverse has already been calculated, then the function should retrieve the inverse from the cache.